backend/requirements.txt
fastapi
uvicorn
python-chess
pymongo
pydantic
